Specific user post

+++ admin/includes/view_all_posts.php
@@ -54,7 +54,7 @@
             <tr>
                 <th><input id="selectAllBoxes" type="checkbox"></th>
                 <th>Id</th>
-                <th>Author</th>
+                <th>User</th>
                 <th>Title</th>
                 <th>Category</th>
                 <th>Status</th>
@@ -70,10 +70,20 @@
         <tbody>
         <?php
             if(isset($_GET['delete'])){
-                $the_post_id =$_GET['delete'];
-                $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
-                $delete_query = mysqli_query($connection, $query);
-                confirmQuery($delete_query);
+
+                if(isset($_SESSION['user_role'])){
+
+                    if($_SESSION['user_role'] == 'admin'){
+
+                        $the_post_id =$_GET['delete'];
+                        $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
+                        $delete_query = mysqli_query($connection, $query);
+                        confirmQuery($delete_query);
+                    }
+
+                }
+
+             
             }
         ?>
 
@@ -84,6 +94,7 @@
             while($row = mysqli_fetch_assoc($select_posts)){
                 $post_id = $row['post_id'];
                 $post_author = $row['post_author'];
+                $post_user = $row['post_user'];
                 $post_title = $row['post_title'];
                 $post_category_id = $row['post_category_id'];
                 $post_status = $row['post_status'];
@@ -97,7 +108,24 @@
                 <td><input class='checkBoxes' type='checkbox' name='checkBoxArray[]' value='<?php echo $post_id; ?>'></td>
         <?php
                 echo "<td>$post_id</td>";
-                echo "<td>$post_author</td>";
+
+
+
+                // if(!empty($post_author)){
+                //     echo "<td>$post_author</td>";
+                // }
+                // else{
+                    echo "<td>$post_user</td>";
+                // }
+                
+
+
+
+
+
+
+
+
                 echo "<td>$post_title</td>";
 
                 $query = "SELECT * FROM categories WHERE cat_id = $post_category_id";
